+++ src/core/game/PlayerImpl.ts
@@ -537,12 +537,12 @@ export class PlayerImpl implements Player {
     this.sentDonations.push(new Donation(recipient, this.mg.ticks()));
     recipient.addGold(this.removeGold(gold));
     this.mg.displayMessage(
-      `Sent ${renderNumber(gold)} golds to ${recipient.name()}`,
+      `Sent ${renderNumber(gold)} gold to ${recipient.name()}`,
       MessageType.INFO,
       this.id(),
     );
     this.mg.displayMessage(
-      `Recieved ${renderNumber(gold)} golds from ${this.name()}`,
+      `Recieved ${renderNumber(gold)} gold from ${this.name()}`,
       MessageType.SUCCESS,
       recipient.id(),
     );
@@ -599,13 +599,12 @@ export class PlayerImpl implements Player {
   }
 
   removeGold(toRemove: Gold): number {
-    if (toRemove > this._gold) {
-      throw Error(
-        `Player ${this} does not enough gold (${toRemove} vs ${this._gold}`,
-      );
+    if (toRemove <= 1) {
+      return 0;
     }
-    this._gold -= toInt(toRemove);
-    return Number(toRemove);
+    const actualRemoved = minInt(this._gold, toInt(toRemove));
+    this._gold -= actualRemoved;
+    return Number(actualRemoved);
   }
 
   population(): number {
